Build and export JS and *.d.ts files (#12)
Attempt to try to export *.js and *.d.ts files so library can be
consumed.

It doesn't appear to fix my issue unfortunately, but still might be a
good thing to do? It's possible Copilot was misleading me here.

I think if we do this we do need to add a build step somehow.