Add access to a `callInvoker` (#22)
According to [The Big O of Code
Reviews](https://www.egorand.dev/the-big-o-of-code-reviews/), this is a
O(_n_) change.

This PR gets a `callInvoker` and feeds it to the generated bindings.

The `callInvoker` comes from one of three places:

- Android and iOS get it from the app. The turbo-module templates have
been updated and tested against the target library's example app.
- A stub for testing.

The stub has been copied from (and lightly adapted) from the jni source
code. At the moment, the callInvoker is unused, and so the stub has
empty implementations of the key methods.