Prefer interface types to concrete class names
When defining a new class, uniffi-bindgen-react-native routinely
generates an interface.

By convention, we're calling this `FooInterface`.

```ts
interface FooInterface {
    method(obj: Bar): Baz
}

class Foo implements FooInterface {
    method(obj: Bar): Baz
}
```

When a method on a `FooInterface` returns another object, or accepts an
object as an argument— say `Bar`— prior to this commit, both the `FooInterface` and
the concrete `Foo` type would use the concrete `Bar`.

```ts
interface FooInterface {
    method(obj: Bar): Baz
}

class Foo implements FooInterface {
    method(obj: Bar): Baz {
        // …
    }
}
```

This changes this, to use the interface name.

```ts
interface FooInterface {
    method(obj: BarInterface): BazInterface
}

class Foo implements FooInterface {
    method(obj: BarInterface): BazInterface {
        // …
    }
}
```

This makes it easy for creating mocks and dummy interfaces.