Eliminate External type from rendering by adding another pass
Before rendering any files, a pass is taken of all ComponentInterfaces.

Each type is collected, and added to a type map, keyed by module-path and name.

When an external type is encountered later, the type map is consulted, and the actual type
is used.

This is to remove a deficiency in `Type::External`: the `ExternalKind` enum did not map completely onto a `Type`:
before this change, it wasn't clear what sort of type was being imported, and so it was difficult/impossible
to work out __how__ to import (`import { X }` vs `import { type X  }`), or __what__ to import: `X` vs `XInterface`.