Deterministic code generation order (#87)
According to [The Big O of Code
Reviews](https://www.egorand.dev/the-big-o-of-code-reviews/), this is a
O(_n_) change.

This PR adds determinism to the generation of the code. This allows
generated code to be checked into version control and it not be a hot
mess each time a change is made.