Hide private fields in tagged enums and errors (#84)
According to [The Big O of Code
Reviews](https://www.egorand.dev/the-big-o-of-code-reviews/), this is a
O(_m * n_) change.

Potentially fixes #75.

This PR does a couple of things for TaggedEnums:

- we eliminate the `Variant__data` intermediate type and replace it with
a macro to render the type explicitly. This is to make the values and
their types show up in autocomplete.
- we move or remove `__uniffiTypeName`, `__variant` and `__variantName`
from `UniffiEnum` and `UniffiError`. I think that this fixes the bug
#75. If not, I don't think it's harmful to keep this fix.
- we introduce a `playground` directory into the `typescript/tests`
directory: this is explicitly for experimenting with hand writing
generated code before touching the templates.