Detect if ArrayBuffer is really a typed array
Fixes #91.

Typescript accepts a typed array (`Uint32Array`, and friends) where it expects an `ArrayBuffer`.

This could mean passing an `Uint32Array` slice where a `Vec<u8>` is expected.

This commit ensures the `FfiConverterArrayBuffer` handles this correctly: it now sends the slice of the underlying buffer that the typed array represents.

Unfortunately, it is not possible to detect pathological cases:

```ts
const ab = new ArrayBuffer(256);
const ta8 = new Uint8Array(ab, 0, 64); // view of `ab`.
const ta32 = new Uint32Array(ta8, 0, 16); // nested view of view of `ab`.
````

I would like the `ffi-converter` to detect this case, but it can't.

This is because the constructor of `Uint32Array` treats the value of `ta8` as an `ArrayLike` iterable, rather than an `ArrayBufferView` or `ArrayBuffer`.

This should likely be documented.