Build for selected Android targets
Fixes #106

This issue is conceptually simple: we need to add one line to the `build.gradle` template.

Driving that line in the file is data about the targets being built.

`targets` comes from the `ProjectConfig` (a.k.a. `ubrn.config.yaml`), which makes this straightforward.

HOWEVER, `--targets` can be used when building!

SO, we can mutate the `ProjectConfig` while we're building.

BUT we don't want the second command overwriting the `build.gradle` so carefully specified in the first.

```sh
ubrn build android --and-generate --config ubrn.config.yaml --targets aarch64-linux-android,armv7-linux-androideabi

ubrn build ios --and-generate --config ubrn.config.yaml
```

THIS MEANT: adding platform specificity to the each of the generated files.