Default to x86_64-apple-ios if ubrn is built for x86 (#135)
This adds `x86_64-apple-ios` rather than `aarch64-apple-ios-sim` to the
default targets on iOS if UBRN itself was built for x86.

Given that x86 Macs are slowly becoming extinct, I think, it may not be
worth to include both targets by default because it'll result in a lot
of pointless compilation.

I believe the `ARCH` value is baked in at compile time. So this may not
work correctly in some edge cases. @trevoranderson, maybe you could give
this a test on your Intel Mac?

Fixes: #134