Update for react-native 0.76 (#147)
This PR tests the React Native update to 0.76 and fixes breakages
introduced.

Breaking changes:

- We change [the `CMakeLists.txt` to use
`libreactnative.so`](https://github.com/react-native-community/discussions-and-proposals/discussions/816).
Currently, we're using `externalNativeBuild`, so we're not getting the
benefit of the Kotlin gradle plugin to set `REACTNATIVE_MERGED_SO` so we
switch on the React Native version.
- The minimum Xcode version increased from `13.0` to `15.1`. This meant
upgrading Xcode on my machine, which found some fragility in the `xcrun
simctl` and jq scripting in the tests.
- The minimum Android SDK version increased. If we commit to supporting
0.75.4 at the same time as 0.76, then I updated the version passed to
`cargo ndk` to what 0.75.4 supports, i.e. 23. This was an increase from
21.
- added documentation in the getting started guide about what version
react native is.