Reproduce callback deadlock
A minimal test which exhibits a deadlock.

On my machine, I can get it to deadlock about 25-35% of the time.

This is with the script:

```sh
fixture=fixtures/callbacks-regression
for i in {1..50} ; do
  echo
  echo "Starting $i"
  cargo xtask run \
    --crate $fixture \
    --ts-dir $fixture/generated \
    --cpp-dir $fixture/generated/cpp \
    $fixture/tests/bindings/test_callbacks_regression.ts
done
```

Counting the incidence of `timed out` in the output and multiplying by 2.

Increasing the work done in the callback (e.g. by uncommenting the `console.log`) gave this error. I'm fairly sure that this is a separate error, and likely an artifact of the test-harness.

```sh
-- Going to 1024, now at: 264
-- Going to 1024, now at: 265
Assertion failed: (usedDbg_ && "Stats not submitted."), function ~CollectionStats, file HadesGC.cpp, line 259.
```

This error is the reason why I can't make a test case to reproduce on CI.

For now, I'm going to treat this error as different to the deadlock.