Fixup callbacks deadlock when calling callback very frequently (#158)
Fixes #157

Minimal test: 146cd396aed959e69bf78d898986656d67675d92

The change here is in `UniffiCallInvoker`, reverting the implementation
of `invokeBlocking` back to the Promise based implementation introduced
in PR https://github.com/jhugman/uniffi-bindgen-react-native/pull/88.

The deadlock rate for the test is now 0/50.

/cc @tayrevor

Adding any more to the implementation (e.g. uncommenting the
`console.log`) causes the same error as before, but it is completely
deterministic:

```sh
-- Going to 1024, now at: 264
-- Going to 1024, now at: 265
Assertion failed: (usedDbg_ && "Stats not submitted."), function ~CollectionStats, file HadesGC.cpp, line 259.
```

On my machine it _always_ stops after 265/1024. I am still unsure if
this is a problem with production uniff-bindgen-react-native code, an
artifact of the test-harness or something weirder.

I'm going to suggest that this is addressed in a separate issue.