Make first run more informative while compiling (#185)
Fixes #181 

According to [The Big O of Code
Reviews](https://www.egorand.dev/the-big-o-of-code-reviews/), this is a
O(_n_) change.

This uses the "naive" solution suggested by @kraenhansen—it just detects
the absence of the `target` directory.

I added the `--quiet` flag sometime ago because the default cargo output
of the `ubrn` was getting to be annoying. Hopefully this is a reasonable
compromise.