Switch from ArrayBuffer to ByteArray (#187)
According to [The Big O of Code
Reviews](https://www.egorand.dev/the-big-o-of-code-reviews/), this is a
O(_n_) change.

ubrn lowers most types to an `ArrayBuffer` and then maps it to a
`jsi::ArrayBuffer`, and then on to `RustBuffer`.
`wasm-bindgen` maps a `Uint8Array` on to a `Vec<u8>`.

In order to lower types to a `Uint8Array` for WASM and `ArrayBuffer` for
JSI, some amount of bundler violence was needed. Given we're using
multiple bundlers to package up the frontend Typescript, _and_ we're not
really in control of what bundler the developer uses, this doesn't seem
like a viable approach.

Instead, this PR changes ubrn to lower types down to a `Uint8Array`, so
that both WASM and React Native use the same underlying representation
of byte arrays.

Unfortunately, jsi doesn't expose API for typed arrays (it may in the
future), so we look up the [`buffer`
property](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/TypedArray#instance_properties)
to get an underlying `ArrayBuffer`.

On the way back, an object of type `{ buffer: ArrayBuffer; }` is
returned, which is enough to make the uniffi work and the typescript
compiler think it that a typed array has come back from Rust/C++.

Most of this PR was mechanical refactoring (e.g. `ArrayBuffer` renamed
to `UniffiByteArray`) but there are enough manual changes which merit it
being O(_n_).