Stabilise `require.resolve` by looking up `package.json` instead of entrypoint
This change decouples the path lookup from the internal entrypoint of `uniffi-bindgen-react-native`.

Since `package.json` always lives at the root of an npm module, `require.resolve`-ing it and then either stripping `/package.json` from the end, or passing the path to `path.dirname()` will get the module's path on-disk.