Use the C-ABI to call into Rust crates for WASM! (#202)
This PR started off as getting errors working.

This, turned out, fairly straightforward. However, showing that the
fixed worked by enabling a test fixture turned out to be. less
straightforward.

This led to a re-write of the rust generation. In this PR, we switch
from using a Rust-to-Rust calls which happen to be exposed by uniffi as
`extern "C"` to relying on the ABI and only the ABI.

This means we do not have to rely on Rust visibility rules, meaning that
a large number of fixtures now Just Work™, without needing to wait for
uniffi-rs to release `v0.29.x`.

We _don't_ need to keep track of the module paths, so I am hoping that
multi-module crates should be straightforward. The only reason why the
`ext-types` fixture is not enabled is because the `uniffi_one.udl` has a
callback interface.

I don't think this approach is compatible with a some of the
optimizations outlined in
https://github.com/mozilla/uniffi-rs/issues/2331, but I think this is a
good approach for 0.28.