Fixup GC'ing objects with callbacks intermittent crasher (#208)
According to [The Big O of Code
Reviews](https://www.egorand.dev/the-big-o-of-code-reviews/), this is a
O(_n_) change.

Fixes https://github.com/zzorba/ubrn-sample-app/pull/1

The investigation for this was in
https://github.com/zzorba/ubrn-sample-app/pull/1.

The key insight here is that the "fake GC" destructor is calling into
Rust which then calls a free callback in hermes, which somehow triggers
another GC.

The fix is to schedule the free callback on the event loop.

This also fixes previously encountered `"Stats not submitted."`
assertion failures.

Possible other things to do for this: document that you shouldn't call
callbacks in a `Drop::drop` though I'm not sure how to phrase that
without being too restrictive.