Fixup hermes build failures (#217)
This PR changes the version of Hermes used to run the JSI tests to be
the version release in React Native 0.77.

It switches from `main`, which was failing to build on Linux with the
toolchain currently installed on CI.

Subsequent PRs may either change this back and fix the toolchain, or
increment the release branch used.