Ext-types fixture working with WASM! (#226)
According to [The Big O of Code
Reviews](https://www.egorand.dev/the-big-o-of-code-reviews/), this is a
O(_n_) change.

This small PR adds multi-crate support.

The target crate can generate C-ABI anywhere in its module structure,
even within sub-crates. In the generated crate, we are reproducing the
C-ABI per module.

The linking happens at the top level "entrypoint" `lib.rs` file, which
`use`s the crate's exported library.

Most of this PR is just feeding the Crate Metadata into the entrypoint
generation code.