Replace the static local of the callback interface table with a "registry"
The static local was needed to store a vtable, so that a pointer could be
given to Rust.

This had a side effect of being persisted after the end of the jsi::Runtime
and the beginning of a new one, which meant hot-reloading of typescript
didn't work if callback interfaces or foreign traits were used.

I tried to add the vtable to the jsi::Object version of the vtable, but this
destructed too early, meaning that when Rust dropped the callback interface
at the end of a run, it segfaulted.

A partial solution was storing the vtable in a static vector, and then clearing
the vector at the registration of the module (or construction). This worked fine,
but there is a potential race condition: if a callback is called after the
vector is cleared, but before a new one is set, then there would be a crash.

Finally, we come to what we have now: a map, where the registry is not cleared
all at once, but the vtables are replaced one at a time. The time where a dangling
pointer is possible is reduced to nothing.

The side-effect on calling a callback when there is no jsi::Runtime is lost,
but not a segfault.