Typecheck Typescript when generating for WASM (#264)
This fixes a number of type checking errors in the generated typescript
bindings.

One thing we cannot workaround (though is done in the tests), is the
existence of the `FinalizationRegistry`.

This is available in `ES2021` and on, so adding this to the
`compileOptions`/`target` in the libraries `tsconfig.json` is necessary.