Upgrade to work with Typescript 5.7+ (#271)
Fixes #270 .

This bumps the version of typescript to ES2024, 5.8.3.

The most significant change is around the usage of `ArrayBuffer`:

## ⚠️ Breaking change ⚠️ 

Prior to 5.7, Typescript accepted TypedArrays (`Uint8Array`, and
friends) as `ArrayBuffer`: i.e. you could declare a variable as
`ArrayBuffer`, but assign it a `Int32Array`.

Since 5.7, these are distinct types: `ArrayBuffer` and
`ArrayBufferView`, which is the union of all typed arrays and
`DataView`.

This PR removes the unpacking of an `ArrayBufferView` into an
`ArrayBuffer`, because now this code path is never reached: because the
codepath to get there wouldn't compile.