Set page size for android (#294)
Google is soon mandating that all apps targeting Android SDK 15+ have
alignment on 16KB boundaries:
https://developer.android.com/guide/practices/page-sizes

This will be a hard requirement for new submissions in 45 days.  

AFAIK, the impact on extra alignment on older targets should be next to
nothing, so I think we are probably wise to just enable this as a new
default on all android platforms, but I'm open to suggestions.