Stop checking for jsRootDir & friends (#310)
Once https://github.com/jhugman/uniffi-bindgen-react-native/pull/309 has
landed, this should "fix" the Compat tests. I'm not sure if this is a
good fix though. Apparently CRNL doesn't use `jsRootDir` anymore when
generating libraries which means we cannot really check for it remaining
untouched anymore. As a corollary, this may mean that we shouldn't use
`jsRootDir` ourselves either when generating `build.gradle`. I'm not
aware of this causing any issues, however, so just dropping the check
might be ok-ish for now?

Edit: Apparently the `libraryName` and `codegenJavaPackageName`
properties cause the same problem making this even less of a great fix.
🫤